// src/permission/enum/permission.enum.ts
export enum PermissionEnum {
  CREATE_USER = 'CREATE_USER',
  UPDATE_USER = 'UPDATE_USER',
  VIEW_USER = 'VIEW_USER',
  DELETE_USER = 'DELETE_USER',

  CREATE_ORDER = 'CREATE_ORDER',
  UPDATE_ORDER = 'UPDATE_ORDER',
  VIEW_ORDER = 'VIEW_ORDER',
  UPDATE_ORDER_STATUS = 'UPDATE_ORDER_STATUS',
  UPDATE_ORDER_CUSTOMER_DETAILS = 'UPDATE_ORDER_CUSTOMER_DETAILS',

  CREATE_PRODUCT = 'CREATE_PRODUCT',
  UPDATE_PRODUCT = 'UPDATE_PRODUCT',
  VIEW_PRODUCT = 'VIEW_PRODUCT',
  DELETE_PRODUCT = 'DELETE_PRODUCT',

  CREATE_BUCKET = 'CREATE_BUCKET',
  UPDATE_BUCKET = 'UPDATE_BUCKET',
  VIEW_BUCKET = 'VIEW_BUCKET',
  DELETE_BUCKET = 'DELETE_BUCKET',

  CREATE_PACKAGE = 'CREATE_PACKAGE',
  UPDATE_PACKAGE = 'UPDATE_PACKAGE',
  VIEW_PACKAGE = 'VIEW_PACKAGE',
  DELETE_PACKAGE = 'DELETE_PACKAGE',
  MANAGE_PACKAGE_ADDON = 'MANAGE_PACKAGE_ADDON',

  VIEW_CATEGORY = 'VIEW_CATEGORY',
  CREATE_CATEGORY = 'CREATE_CATEGORY',
  UPDATE_CATEGORY = 'UPDATE_CATEGORY',
  DELETE_CATEGORY = 'DELETE_CATEGORY',

  VIEW_OUTLET = 'VIEW_OUTLET',
  CREATE_OUTLET = 'CREATE_OUTLET',
  UPDATE_OUTLET = 'UPDATE_OUTLET',
  DELETE_OUTLET = 'DELETE_OUTLET',

  VIEW_CUTTING_METHOD = 'VIEW_CUTTING_METHOD',
  CREATE_CUTTING_METHOD = 'CREATE_CUTTING_METHOD',
  UPDATE_CUTTING_METHOD = 'UPDATE_CUTTING_METHOD',
  DELETE_CUTTING_METHOD = 'DELETE_CUTTING_METHOD',

  VIEW_DELIVERY_SLOT = 'VIEW_DELIVERY_SLOT',
  CREATE_DELIVERY_SLOT = 'CREATE_DELIVERY_SLOT',
  UPDATE_DELIVERY_SLOT = 'UPDATE_DELIVERY_SLOT',
  DELETE_DELIVERY_SLOT = 'DELETE_DELIVERY_SLOT',

  CREATE_SUB_PRODUCT = 'CREATE_SUB_PRODUCT',
  UPDATE_SUB_PRODUCT = 'UPDATE_SUB_PRODUCT',
  VIEW_SUB_PRODUCT = 'VIEW_SUB_PRODUCT',
  DELETE_SUB_PRODUCT = 'DELETE_SUB_PRODUCT',

  SUB_PRODUCT_PRICE_MANAGE = 'SUB_PRODUCT_PRICE_MANAGE',
  PACKAGE_PRICE_MANAGE = 'PACKAGE_PRICE_MANAGE',

  VIEW_BATCH = 'VIEW_BATCH',
  CREATE_BATCH = 'CREATE_BATCH',
  UPDATE_BATCH = 'UPDATE_BATCH',
  DELETE_BATCH = 'DELETE_BATCH',
  ADD_PURCHASE_ITEMS = 'ADD_PURCHASE_ITEMS',
  UPDATE_PURCHASE_ITEMS = 'UPDATE_PURCHASE_ITEMS',

  VIEW_STOCK = 'VIEW_STOCK',
  CREATE_STOCK = 'CREATE_STOCK',
  UPDATE_STOCK = 'UPDATE_STOCK',

  CREATE_SUPPLIER = 'CREATE_SUPPLIER',
  UPDATE_SUPPLIER = 'UPDATE_SUPPLIER',
  VIEW_SUPPLIER = 'VIEW_SUPPLIER',
  DELETE_SUPPLIER = 'DELETE_SUPPLIER',

  VIEW_DRIVER = 'VIEW_DRIVER',
  CREATE_DRIVER = 'CREATE_DRIVER',
  UPDATE_DRIVER = 'UPDATE_DRIVER',
  DELETE_DRIVER = 'DELETE_DRIVER',

  VIEW_CUSTOMER = 'VIEW_CUSTOMER',
  TERMINATE_CUSTOMER = 'TERMINATE_CUSTOMER',

  CREATE_ROLE = 'CREATE_ROLE',
  UPDATE_ROLE = 'UPDATE_ROLE',
  VIEW_ROLE = 'VIEW_ROLE',
  DELETE_ROLE = 'DELETE_ROLE',
  ASSIGN_PERMISSION = 'ASSIGN_PERMISSION',

  VIEW_PAYMENTS = 'VIEW_PAYMENTS',

  MANUAL_ORDER = 'MANUAL_ORDER',

  PERMISSION_MANAGE = 'PERMISSION_MANAGE',

  CREATE_PAYMENTS = 'CREATE_PAYMENTS',
  VIEW_REPORT = 'VIEW_REPORT',
}